Fixed some issues with the way we handle preserveAspectRation.  The old code basically didn't allow you to not do viewBox to viewport scaling. We were always defaulting to someing (ie. XMidYMid meet). Once I fixed that, I had to do a bit of bug fixing because other parts of the code weren't expecting null values for AspectRatioAlignment and AspectRatioScale.
One improvement as a result of this change is that the ensureRootBox() method we removed a few patches ago, is now replaced by an automatic process that does the same thing. If you specify and alignment when calling renderToX(), the code assumes you therefore want to fill the viewport, and handles it appropriately.